feat: Add posts/following/followers update on respective actions

+++ src/controllers/follows.controller.ts
@@ -1,5 +1,6 @@
 import { Request, Response } from "express";
 import Follow, { FollowInput } from "../models/follow.model";
+import User from "../models/user.model";
 
 async function getCurrentUserFollow(req: Request, res: Response) {
   const { targetId } = req.params;
@@ -62,6 +63,9 @@ async function follow(req: Request, res: Response) {
   };
   const created = await Follow.create(followInput);
 
+  await User.findOneAndUpdate({ _id: user._id }, { $inc: { following: 1 } });
+  await User.findOneAndUpdate({ _id: targetId }, { $inc: { followers: 1 } });
+
   return res.status(201).json(created);
 }
 
@@ -69,7 +73,17 @@ async function unfollow(req: Request, res: Response) {
   const { targetId } = req.params;
   const user = req.user;
 
+  if (!targetId) {
+    return res.status(422).json({ message: "No follow target specified" });
+  }
+  if (!user) {
+    return res.status(401).json({ message: "Could not identify user" });
+  }
+
   await Follow.findOneAndDelete({ follower: user?._id, target: targetId });
+
+  await User.findOneAndUpdate({ _id: user._id }, { $dec: { following: -1 } });
+  await User.findOneAndUpdate({ _id: targetId }, { $inc: { followers: -1 } });
 
   return res.status(200).json({ message: "Post deleted successfully." });
 }
